-(xx__freebsd2.c) - fix size_t to build on FreeBSD 2.2.x
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2740 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/ucd-snmp/memory_freebsd2.c
@@ -214,9 +214,9 @@ void swapmode(void)
 
 unsigned char *var_extensible_mem(struct variable *vp,
 				  oid *name,
-				  int *length,
+				  size_t *length,
 				  int exact,
-				  int *var_len,
+				  size_t *var_len,
 				  WriteMethod **write_method)
 {
     static long long_ret;
